refactor: move generic map data code into separate files
The MapShortcode class contained a lot of code that wasn’t technically
required for rendering the shortcode. The logic has been moved into
separate classes so that it is easier to re-use it.
