+++ includes/Public.php
@@ -1,6 +1,6 @@
 <?php
 
-use CommonsBooking\Map\MapShortcode;
+use CommonsBooking\Map\MapData;
 use CommonsBooking\Migration\Migration;
 use CommonsBooking\View\Booking;
 use CommonsBooking\View\Calendar;
@@ -165,10 +165,10 @@ if ( is_admin() ) {
 }
 
 // Map ajax
-add_action( 'wp_ajax_cb_map_locations', array( MapShortcode::class, 'get_locations' ) );
-add_action( 'wp_ajax_nopriv_cb_map_locations', array( MapShortcode::class, 'get_locations' ) );
-add_action( 'wp_ajax_cb_map_geo_search', array( MapShortcode::class, 'geo_search' ) );
-add_action( 'wp_ajax_nopriv_cb_map_geo_search', array( MapShortcode::class, 'geo_search' ) );
+add_action( 'wp_ajax_cb_map_locations', array( MapData::class, 'get_locations' ) );
+add_action( 'wp_ajax_nopriv_cb_map_locations', array( MapData::class, 'get_locations' ) );
+add_action( 'wp_ajax_cb_map_geo_search', array( MapData::class, 'geo_search' ) );
+add_action( 'wp_ajax_nopriv_cb_map_geo_search', array( MapData::class, 'geo_search' ) );
 
 // Query vars
 function commonsbooking_query_vars( $qvars ) {
